fix(kube/apps): make filter go down
