+++ app/react/components/datatables/Datatable.tsx
@@ -14,7 +14,7 @@ import {
   getExpandedRowModel,
   TableOptions,
 } from '@tanstack/react-table';
-import { ReactNode, useMemo } from 'react';
+import { ComponentProps, ReactNode, useMemo } from 'react';
 import clsx from 'clsx';
 import _ from 'lodash';
 
@@ -50,6 +50,10 @@ export type PaginationProps =
 export interface Props<D extends DefaultType> extends AutomationTestingProps {
   dataset: D[];
   columns: TableOptions<D>['columns'];
+  /**
+   * can be used to customize the sidebar. receives the search bar, table actions and table settings elements.
+   */
+  renderSidebar?: ComponentProps<typeof DatatableHeader>['renderSidebar'];
   renderTableSettings?(instance: TableInstance<D>): ReactNode;
   renderTableActions?(selectedRows: D[]): ReactNode;
   disableSelect?: boolean;
@@ -67,11 +71,6 @@ export interface Props<D extends DefaultType> extends AutomationTestingProps {
   getRowCanExpand?(row: Row<D>): boolean;
   noWidget?: boolean;
   extendTableOptions?: (options: TableOptions<D>) => TableOptions<D>;
-  renderHeaderRightSide?(
-    searchBar: ReactNode,
-    tableActions: ReactNode,
-    tableTitleSettings: ReactNode
-  ): ReactNode;
 }
 
 export function Datatable<D extends DefaultType>({
@@ -99,7 +98,7 @@ export function Datatable<D extends DefaultType>({
   totalCount = dataset.length,
   isServerSidePagination = false,
   extendTableOptions = (value) => value,
-  renderHeaderRightSide,
+  renderSidebar,
 }: Props<D> & PaginationProps) {
   const pageCount = useMemo(
     () => Math.ceil(totalCount / settings.pageSize),
@@ -182,7 +181,7 @@ export function Datatable<D extends DefaultType>({
         description={description}
         renderTableActions={() => renderTableActions(selectedItems)}
         renderTableSettings={() => renderTableSettings(tableInstance)}
-        renderRightSide={renderHeaderRightSide}
+        renderSidebar={renderSidebar}
       />
 
       <DatatableContent<D>
