+++ app/react/kubernetes/applications/ListView/ApplicationsStacksDatatable/types.ts
@@ -9,3 +9,10 @@ export interface TableSettings
   extends BasicTableSettings,
     RefreshableTableSettings,
     SystemResourcesTableSettings {}
+
+export interface Namespace {
+  Id: string;
+  Name: string;
+  Yaml: string;
+  IsSystem?: boolean;
+}
